requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
appdirs==1.4.4
beautifulsoup4==4.10.0; python_version >= '3.1'
bs4==0.0.1
cachetools==5.0.0; python_version ~= '3.7'
certifi==2021.10.8
charset-normalizer==2.0.12; python_version >= '3'
cssselect==1.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
fake-useragent==0.1.11
google-api-core==2.5.0; python_version >= '3.6'
google-api-python-client==2.37.0
google-auth-httplib2==0.1.0
google-auth==2.6.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
googleapis-common-protos==1.54.0; python_version >= '3.6'
httplib2==0.20.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
idna==3.3; python_version >= '3'
importlib-metadata==4.11.1; python_version >= '3.7'
lxml==4.8.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
parse==1.19.0
praw==7.5.0
prawcore==2.3.0; python_version ~= '3.6'
protobuf==3.19.4; python_version >= '3.5'
pyasn1-modules==0.2.8
pyasn1==0.4.8
pyee==8.2.2
pyparsing==3.0.7; python_version >= '3.1'
pyppeteer==1.0.2; python_version >= '3.7' and python_version < '4'
pyquery==1.4.3
python-dotenv==0.19.2
requests-html==0.10.0
requests==2.27.1
rsa==4.8; python_version >= '3.6'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
soupsieve==2.3.1; python_version >= '3.6'
tqdm==4.62.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
update-checker==0.18.0
uritemplate==4.1.1; python_version >= '3.6'
urllib3==1.26.8; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
w3lib==1.22.0
websocket-client==1.2.3; python_version >= '3.6'
websockets==10.1; python_version >= '3.7'
zipp==3.7.0; python_version >= '3.7'
